Printing Area of Circle
Your task now is to write a Python program which accepts the radius of a circle from the user and computes area.
Your Input and Output should look something like this
Input the radius of the circle : 1.1 The area of the circle with radius 1.1 is: 3.8013271108436504

+++ TASK 1.py
@@ -4,26 +4,7 @@
 # In[11]:
 
 
-r=int(input("Enter radius of circle:"))
+r=float(input("Enter radius of circle:"))
 pi=3.141592653589793238
 area=pi*r*r
 print("Area of Circle is",area)
-
-
-# In[ ]:
-
-
-
-
-
-# In[ ]:
-
-
-
-
-
-# In[ ]:
-
-
-
-
